@@ -1,5 +1,5 @@
 {
-  "name": "qgnioqqrg.github",
+  "name": "clean-code",
   "version": "0.0.0",
   "private": true,
   "scripts": {
@@ -11,29 +11,29 @@
     "e2e": "ng e2e"
   },
   "dependencies": {
-    "@angular/animations": "^15.1.2",
-    "@angular/common": "^15.1.2",
-    "@angular/compiler": "^15.1.2",
-    "@angular/core": "^15.1.2",
-    "@angular/forms": "^15.1.2",
-    "@angular/platform-browser": "^15.1.2",
-    "@angular/platform-browser-dynamic": "^15.1.2",
-    "@angular/router": "^15.1.2",
-    "rxjs": "^7.8.0",
-    "tslib": "^2.5.0",
-    "zone.js": "^0.12.0"
+    "@angular/animations": "^15.2.0",
+    "@angular/common": "^15.2.0",
+    "@angular/compiler": "^15.2.0",
+    "@angular/core": "^15.2.0",
+    "@angular/forms": "^15.2.0",
+    "@angular/platform-browser": "^15.2.0",
+    "@angular/platform-browser-dynamic": "^15.2.0",
+    "@angular/router": "^15.2.0",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.12.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "15.1.0-next.2",
-    "@angular/cli": "15.1.0-next.2",
-    "@angular/compiler-cli": "15.1.0-next.2",
-    "@types/jasmine": "~4.0.0",
-    "jasmine-core": "~4.1.0",
-    "karma": "~6.3.0",
+    "@angular-devkit/build-angular": "^15.2.5",
+    "@angular/cli": "~15.2.5",
+    "@angular/compiler-cli": "^15.2.0",
+    "@types/jasmine": "~4.3.0",
+    "jasmine-core": "~4.5.0",
+    "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.2.0",
-    "karma-jasmine": "~5.0.0",
-    "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.7.2"
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.0.0",
+    "typescript": "~4.9.4"
   }
 }
