@@ -1,4 +1,4 @@
-export enum VersionType {
+export enum VERSION_TYPE {
   VIDEO = 'video',
   IMAGE = 'image',
 }
