@@ -1,10 +1,10 @@
 import { UniqueId } from './unique-id';
-import { VersionType } from './version-type';
+import { VERSION_TYPE } from './version-type';
 import { VersionUrl } from './version-url';
 
 export type Version = {
   id: UniqueId;
-  type: VersionType;
+  type: VERSION_TYPE;
   url: VersionUrl;
   name: string;
 };
